# docs/swagger.yaml
basePath: /api/v1
definitions:
  apiutil.errorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  apiutil.errorResponse:
    properties:
      error:
        $ref: '#/definitions/apiutil.errorBody'
    type: object
  handlers.CreateStudentInput:
    properties:
      department_id:
        type: integer
      email:
        type: string
      faculty_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - faculty_id
    - first_name
    - last_name
    - password
    type: object
...
  handlers.TransferStudentInput:
    properties:
      department_id:
        type: integer
      faculty_id:
        type: integer
    required:
    - faculty_id
    type: object
  handlers.UpdateStudentInput:
    properties:
      department_id:
        type: integer
      faculty_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
    required:
    - faculty_id
    - first_name
    - last_name
    type: object
...
  models.Role:
    enum:
    - admin
    - employee
    - professor
    - student
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleEmployee
    - RoleProfessor
    - RoleStudent
...
  models.StudentDepartmentHistory:
    properties:
      created_at:
        type: string
      from_department_id:
        type: integer
      from_faculty_id:
        type: integer
      id:
        type: integer
      student:
        $ref: '#/definitions/models.User'
      student_id:
        type: integer
      to_department_id:
        type: integer
      to_faculty_id:
        type: integer
      transferred_at:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      department_id:
        type: integer
      department_transfers:
        items:
          $ref: '#/definitions/models.StudentDepartmentHistory'
        type: array
      email:
        type: string
      faculty_id:
        type: integer
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      national_id:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the UniOne Go backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: UniOne API
  version: "1.0"
paths:
...
  /admin/students:
    get:
      consumes:
      - application/json
      description: Returns a list of all students.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerAuth: []
      summary: List all students
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Creates a new student user and profile.
      parameters:
      - description: Student details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStudentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiutil.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a new student
      tags:
      - admin
  /admin/students/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a student by their ID.
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a student
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Returns a specific student by their ID.
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apiutil.errorResponse'
      security:
      - BearerAuth: []
      summary: Get student by ID
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates the specified student's information.
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateStudentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiutil.errorResponse'
      security:
      - BearerAuth: []
      summary: Update student details
      tags:
      - admin
  /admin/students/{id}/transfer:
    post:
      consumes:
      - application/json
      description: Record a student transfer and update their current organization
        link.
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transfer details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.TransferStudentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Transfer a student to another faculty/department
      tags:
      - admin
  /admin/students/{id}/transfers:
    get:
      consumes:
      - application/json
      description: Returns the history of organization transfers for a student.
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentDepartmentHistory'
            type: array
      security:
      - BearerAuth: []
      summary: Get student transfer history
      tags:
      - admin
  /admin/students/export:
    get:
      description: Returns an Excel file containing the list of students.
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel binary data
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Export students to Excel
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"